requirements.txt
torch~=2.2.2
transformers~=4.39.3
pm4py~=2.7.11.4
pandas~=2.2.1
matplotlib~=3.8.3
    
numpy~=1.25.2
tqdm~=4.66.2
Pillow~=10.0.0